test runner updates

+++ src/scripts/organizeHome.sh
@@ -2,5 +2,5 @@
 
 source "$(dirname "$0")/utils.sh"
 
-rmdir "$HOME/Templates" 2>>"$ERROR_LOG_FILE" || true
+[[ -d "$HOME/Templates" ]] && rmdir "$HOME/Templates" 2>>"$ERROR_LOG_FILE" || true
 mkdir -p "$HOME/Books" "$HOME/Games" "$HOME/Hacking" "$HOME/Projects" 2>>"$ERROR_LOG_FILE" || true
